+++ inmanta_plugins/podman/resources/abc.py
@@ -29,6 +29,8 @@ class ResourceABC(inmanta.resources.PurgeableResource):
         "owner",
         "name",
     )
+    owner: str
+    name: str
 
     @classmethod
     def get_q(
@@ -39,11 +41,14 @@ class ResourceABC(inmanta.resources.PurgeableResource):
         return f"owner={entity.owner}&name={entity.name}"
 
 
-class HandlerABC(inmanta.agent.handler.CRUDHandler):
+ABC = typing.TypeVar("ABC", bound=ResourceABC)
+
+
+class HandlerABC(inmanta.agent.handler.CRUDHandlerGeneric[ABC]):
     def run_command(
         self,
         ctx: inmanta.agent.handler.HandlerContext,
-        resource: inmanta.resources.Resource,
+        resource: ABC,
         *,
         command: list[str],
         timeout: int,