+++ tests/test_image_from_registry.py
@@ -35,10 +35,9 @@ def test_model(project: Project, purged: bool = False) -> None:
             os=std::linux,
         )
 
-        podman::Network(
+        podman::ImageFromRegistry(
             host=host,
-            name="test-net",
-            subnets=[Subnet(subnet="172.45.0.0/24")],
+            name="docker.io/library/alpine:latest",
             purged={json.dumps(purged)},
         )
     """
@@ -47,13 +46,13 @@ def test_model(project: Project, purged: bool = False) -> None:
 
 
 def test_deploy(project: Project) -> None:
-    # Make sure the network is there
+    # Make sure the busybox image is there
     test_model(project, purged=False)
-    project.deploy_resource("podman::Network")
-    assert not project.dryrun_resource("podman::Network")
+    project.deploy_resource("podman::ImageFromRegistry")
+    assert not project.dryrun_resource("podman::ImageFromRegistry")
 
-    # Make sure the network is gone
+    # Make sure the image is gone
     test_model(project, purged=True)
-    assert project.dryrun_resource("podman::Network")
-    project.deploy_resource("podman::Network")
-    assert not project.dryrun_resource("podman::Network")
+    assert project.dryrun_resource("podman::ImageFromRegistry")
+    project.deploy_resource("podman::ImageFromRegistry")
+    assert not project.dryrun_resource("podman::ImageFromRegistry")