+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## v0.1.0 - ?
 
+- Add support for podman Image
 - Add support for podman Network
 - Add support for systemd container and pod services